# naPO/apps/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: naPO Backend API
  description: |
    naPO (National Policy Observatory) Backend API

    자연어 쿼리를 통해 한국 선거 및 정책 데이터에 접근하는 API입니다.

    ## 주요 기능
    - 자연어 쿼리 파싱 및 실행
    - 공공데이터 API 연동 (선거관리위원회, 정당정책 등)
    - PDF 문서 파싱 (OCR 지원)
    - 웹 크롤링
    - 데이터 내보내기 (CSV, JSON)

  version: 1.0.0
  contact:
    name: naPO Team
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
...
paths:
...
  /crawl:
    post:
      tags:
        - Crawl
      summary: 크롤링 작업 시작
      description: 새로운 웹 크롤링 작업을 시작합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - crawlerType
              properties:
                crawlerType:
                  type: string
                  enum:
                    - custom
                    - nec_library
                options:
                  type: object
                  properties:
                    url:
                      type: string
                      format: uri
                    downloadFiles:
                      type: boolean
      responses:
        '200':
          description: 작업 생성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      jobId:
                        type: string
                      status:
                        type: string
        '400':
          description: 유효성 검증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /crawl/{jobId}:
    get:
      tags:
        - Crawl
      summary: 크롤링 작업 상태
      description: 크롤링 작업의 현재 상태를 조회합니다.
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/CrawlJob'
        '404':
          description: 작업을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request data
...
    CrawlJob:
      type: object
      properties:
        id:
          type: string
        jobType:
          type: string
        status:
          type: string
          enum:
            - PENDING
            - RUNNING
            - COMPLETED
            - FAILED
        resultCount:
          type: integer
        errorMessage:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []